linux_wayland_support: refactor
